Implement history heuristic for search

+++ src/Evaluation.hpp
@@ -10,7 +10,7 @@ namespace Brainiac {
      * @brief Piece weights from white's perspective.
      *
      */
-    constexpr std::array<short, 12> PIECE_WEIGHTS = {
+    constexpr std::array<int, 12> PIECE_WEIGHTS = {
         40,
         10,
         55,
@@ -36,7 +36,7 @@ namespace Brainiac {
      * i.e., starting position is on the bottom 2 rows.
      */
     // clang-format off
-    const short QUEEN_MATRIX[64] = {
+    const int QUEEN_MATRIX[64] = {
         -40, -20, -20, -10, -10, -20, -20, -40, 
         -20, 00, 00, 00, 00, 00, 00, -20,
         -20, 00, 10, 10, 10, 10, 00, -20, 
@@ -46,7 +46,7 @@ namespace Brainiac {
         -20, 00, 10, 00, 00, 00, 00, -20, 
         -40, -20, -20, -10, -10, -20, -20, -40,
     };
-    const short BISHOP_MATRIX[64] = {
+    const int BISHOP_MATRIX[64] = {
         -20, -10, -10, -10, -10, -10, -10, -20, 
         -10, 00, 00, 00, 00, 00, 00, -10,
         -10, 00, 50, 10, 10, 50, 00, -10, 
@@ -56,7 +56,7 @@ namespace Brainiac {
         -10, 50, 00, 00, 00, 00, 50, -10, 
         -20, -10, -50, -10, -10, -50, -10, -20,
     };
-    const short KNIGHT_MATRIX[64] = {
+    const int KNIGHT_MATRIX[64] = {
         -50, -40, -30, -30, -30, -30, -40, -50, 
         -40, -20, 00, 00, 00, 00, -20, -40,
         -30, 00, 10, 15, 15, 10, 00, -30, 
@@ -66,7 +66,7 @@ namespace Brainiac {
         -40, -20, 00, 50, 50, 00, -20, -40, 
         -50, -50, -20, -30, -30, -20, -50, -50,
     };
-    const short ROOK_MATRIX[64] = {
+    const int ROOK_MATRIX[64] = {
         00, 00, 00, 00, 00, 00, 00, 00,  
         10, 20, 20, 20, 20, 20, 20, 10,
         -10, 00, 00, 00, 00, 00, 00, -10, 
@@ -76,7 +76,7 @@ namespace Brainiac {
         -10, 00, 00, 00, 00, 00, 00, -10, 
         -30, 30, 40, 10, 10, 00, 00, -30,
     };
-    const short PAWN_MATRIX[64] = {
+    const int PAWN_MATRIX[64] = {
         00, 00, 00, 00, 00, 00, 00, 00, 
         70, 70, 70, 70, 70, 70, 70, 70,
         10, 10, 20, 30, 30, 20, 10, 10, 
@@ -86,7 +86,7 @@ namespace Brainiac {
         50, 10, 10, -25, -25, 10, 10, 50, 
         00, 00, 00, 00, 00, 00, 00, 00,
     };
-    const short KING_MATRIX[64] = {
+    const int KING_MATRIX[64] = {
         -30, -40, -40, -50, -50, -40, -40, -30, 
         -30, -40, -40, -50, -50, -40, -40, -30,
         -30, -40, -40, -50, -50, -40, -40, -30, 
@@ -102,23 +102,23 @@ namespace Brainiac {
      * @brief Compute the material score from white's perspective.
      *
      * @param board
-     * @return short
+     * @return int
      */
-    short compute_material(const Board &board);
+    int compute_material(const Board &board);
 
     /**
      * @brief Compute the placement score from white's perspective.
      *
      * @param board
-     * @return short
+     * @return int
      */
-    short compute_placement(const Board &board);
+    int compute_placement(const Board &board);
 
     /**
      * @brief Evaluate a position for the current turn.
      *
      * @param pos
-     * @return short
+     * @return int
      */
-    short evaluate(Position &pos);
+    int evaluate(Position &pos);
 } // namespace Brainiac
